+++ R/replace.r
@@ -3,18 +3,19 @@
 #' Vectorised over \code{string}, \code{pattern} and \code{replacement}.
 #'
 #' @inheritParams str_detect
-#' @param pattern,replacement Supply separate pattern and replacement strings
-#'   to vectorise over the patterns. References of the form \code{\1},
-#'   \code{\2} will be replaced with the contents of the respective matched
-#'   group (created by \code{()}) within the pattern.
+#' @param replacement A character vector of replacements. Should be either
+#'   length one, or the same length as \code{string} or \code{pattern}.
+#'   References of the form \code{\1}, \code{\2}, etc will be replaced with
+#'   the contents of the respective matched group (created by \code{()}).
 #'
-#'   For \code{str_replace_all} only, you can perform multiple patterns and
-#'   replacements to each string, by passing a named character to
-#'   \code{pattern}.
+#'   To perform multiple replacements in each element of \code{string},
+#'   pass a named vector (\code{c(pattern1 = replacement1)}) to
+#'   \code{str_replace_all}. Alternatively, pass a function to
+#'   \code{replacement}: it will be called once for each match and its
+#'   return value will be used to replace the match.
 #'
-#'   Alternatively, you can supply use a \code{replacement} function:
-#'   it will be called once for each match and its return value will be
-#'   used to replace the match.
+#'   To replace the complete string with \code{NA}, use
+#'   \code{replacement = NA_character_}.
 #' @return A character vector.
 #' @seealso \code{\link{str_replace_na}} to turn missing values into "NA";
 #'   \code{\link{stri_replace}} for the underlying implementation.
@@ -24,6 +25,7 @@
 #' str_replace(fruits, "[aeiou]", "-")
 #' str_replace_all(fruits, "[aeiou]", "-")
 #' str_replace_all(fruits, "[aeiou]", toupper)
+#' str_replace_all(fruits, "b", NA_character_)
 #'
 #' str_replace(fruits, "([aeiou])", "")
 #' str_replace(fruits, "([aeiou])", "\\1\\1")
@@ -40,10 +42,10 @@
 #' str_replace_all(fruits, c("a", "e", "i"), "-")
 #'
 #' # If you want to apply multiple patterns and replacements to the same
-#' # string, pass a named version to pattern.
+#' # string, pass a named vector to pattern.
 #' fruits %>%
 #'   str_c(collapse = "---") %>%
-#'   str_replace_all(c("one" = 1, "two" = 2, "three" = 3))
+#'   str_replace_all(c("one" = "1", "two" = "2", "three" = "3"))
 #'
 #' # Use a function for more sophisticated replacement. This example
 #' # replaces colour names with their hex values.
@@ -104,11 +106,17 @@ str_replace_all <- function(string, pattern, replacement) {
 }
 
 fix_replacement <- function(x) {
+  if (!is.character(x)) {
+    stop("`replacement` must be a character vector", call. = FALSE)
+  }
+
   vapply(x, fix_replacement_one, character(1), USE.NAMES = FALSE)
 }
 
 fix_replacement_one <- function(x) {
-  escape_dollar <- function(x) if (x == "$") "\\$" else x
+  if (is.na(x)) {
+    return(x)
+  }
 
   chars <- str_split(x, "")[[1]]
   out <- character(length(chars))
@@ -149,6 +157,7 @@ fix_replacement_one <- function(x) {
 #' Turn NA into "NA"
 #'
 #' @inheritParams str_replace
+#' @param replacement A single string.
 #' @export
 #' @examples
 #' str_replace_na(c(NA, "abc", "def"))